+++ wbDefinitionsFO3.pas
@@ -10861,6 +10861,8 @@ begin
   DefineFO3d;
   DefineFO3e;
   DefineFO3f;
+
+  wbHEDRVersion := 0.94;
 end;
 
 end.