+++ wbDefinitionsFO4.pas
@@ -17307,6 +17307,12 @@ begin
   end else
     wbCreationClubContentFileName := 'Fallout4.ccc';
 
+  if wbGameMode = gmFO4VR then
+    wbHEDRVersion := 0.95
+  else begin
+    wbHEDRVersion := 1.0;
+    wbHEDRNextObjectID := 1;
+  end;
 end;
 
 initialization