+++ wbDefinitionsTES3.pas
@@ -4182,6 +4182,8 @@ begin
   wbAddGroupOrder(SNDG);
   wbAddGroupOrder(DIAL);
   wbAddGroupOrder(INFO);
+
+  wbHEDRVersion := 1.30;
 end;
 
 initialization