+++ wbDefinitionsTES4.pas
@@ -5236,6 +5236,8 @@ begin
   wbNexusModsUrl := 'https://www.nexusmods.com/oblivion/mods/11536';
   if wbToolMode = tmLODgen then
     wbNexusModsUrl := 'https://www.nexusmods.com/oblivion/mods/15781';
+
+  wbHEDRVersion := 1.0;
 end;
 
 initialization