+++ wbDefinitionsTES5.pas
@@ -14299,6 +14299,7 @@ begin
 
   end;
 
+  wbHEDRVersion := 1.7;
 end;
 
 initialization