+++ wbInit.pas
@@ -232,7 +232,6 @@ begin
       wbShowFlagEnumValue := Settings.ReadBool('Options', 'ShowFlagEnumValue', wbShowFlagEnumValue);
       wbTrackAllEditorID := Settings.ReadBool('Options', 'TrackAllEditorID', wbTrackAllEditorID);
       wbAllowDirectSave := Settings.ReadBool('Options', 'AllowDirectSave', wbAllowDirectSave);
-      wbExtendedESL := Settings.ReadBool('Options', 'ExtendedESL', wbExtendedESL);
       wbSortINFO := Settings.ReadBool('Options', 'SortINFO', wbSortINFO);
       wbFillPNAM := Settings.ReadBool('Options', 'FillPNAM', wbFillPNAM);
       wbCollapseConditions := Settings.ReadBool('Options', 'CollapseConditions', wbCollapseConditions);
@@ -962,9 +961,6 @@ begin
   end else
     if FindCmdLineSwitch('AllowDirectSaves') then
       wbAllowDirectSave := True;
-
-    if FindCmdLineSwitch('ExtendedESL') then
-      wbExtendedESL := True;
 
   if FindCmdLineSwitch('IKnowWhatImDoing') then
     wbIKnowWhatImDoing := True;