+++ wbInterface.pas
@@ -179,7 +179,9 @@ var
   wbShrinkButtons          : Boolean  = False;
   wbCollapseConditions     : Boolean  = True;
   wbCollapseBenignArray    : Boolean  = True;
-  wbExtendedESL            : Boolean  = False;
+
+  wbHEDRVersion            : Double = 1.0;
+  wbHEDRNextObjectID       : Integer = $800;
 
   wbGlobalModifedGeneration : UInt64;
 
@@ -1265,6 +1267,8 @@ type
     function GetNextObjectID: Cardinal;
     procedure SetNextObjectID(aObjectID: Cardinal);
 
+    function GetAllowHardcodedRangeUse: Boolean;
+
     function HasONAM: Boolean;
     procedure MarkHeaderModified;
 
@@ -1358,6 +1362,9 @@ type
 
     property CompareToFile: IwbFile
       read GetCompareToFile;
+
+    property AllowHardcodedRangeUse: Boolean
+      read GetAllowHardcodedRangeUse;
   end;
 
   IwbDataContainer = interface(IwbContainer)
@@ -12844,9 +12851,16 @@ begin
     _File := aElement._File;
     if Assigned(_File) then begin
       try
-        MainRecord := _File.RecordByFormID[TwbFormID.FromCardinal(aInt), True, aElement.MastersUpdated];
+
+        if (FormID.ObjectID < $800) and not _File.AllowHardcodedRangeUse then begin
+          FormID.FileID := TwbFileID.Null;
+          MainRecord := wbGetGameMasterFile.RecordByFormID[FormID, True, aElement.MastersUpdated];
+        end else
+          MainRecord := _File.RecordByFormID[FormID, True, aElement.MastersUpdated];
+
         if Assigned(MainRecord) then
           Exit;
+
         if wbDisplayLoadOrderFormID then
           FormID := _File.FileFormIDtoLoadOrderFormID(FormID, aElement.MastersUpdated);
       except
@@ -13043,15 +13057,16 @@ begin
   if not wbDisplayLoadOrderFormID then
     Exit;
 
-  if Result < $800 then
-    Exit;
-
   if Result = $FFFFFFFF then
     Exit;
 
   if Assigned(aElement) then begin
     _File := aElement._File;
     if Assigned(_File) then begin
+      if not _File.AllowHardcodedRangeUse then
+        if Result < $800 then
+          Exit;
+
       //Result is a load order FormID right now, we need to store a file specific FormID
       Result := _File.LoadOrderFormIDtoFileFormID(TwbFormID.FromCardinal(Result), aElement.MastersUpdated).ToCardinal;
     end;
@@ -13329,7 +13344,8 @@ end;
 
 function TwbFormIDDefFormater.GetLinksTo(aInt: Int64; const aElement: IwbElement): IwbElement;
 var
-  _File : IwbFile;
+  _File  : IwbFile;
+  FormID : TwbFormID;
 begin
   Result := nil;
 
@@ -13344,20 +13360,33 @@ begin
   if Assigned(aElement) then begin
     _File := aElement._File;
     if Assigned(_File) then try
-      Result := _File.RecordByFormID[TwbFormID.FromCardinal(aInt), True, aElement.MastersUpdated];
+      FormID := TwbFormID.FromCardinal(aInt);
+      if (FormID.ObjectID < $800) and not _File.AllowHardcodedRangeUse then begin
+        FormID.FileID := TwbFileID.Null;
+        Result := wbGetGameMasterFile.RecordByFormID[FormID, True, aElement.MastersUpdated];
+      end else
+        Result := _File.RecordByFormID[FormID, True, aElement.MastersUpdated];
     except end;
   end;
 end;
 
 function TwbFormIDDefFormater.GetMainRecord(aInt: Int64; const aElement: IwbElement): IwbMainRecord;
 var
-  _File: IwbFile;
+  _File  : IwbFile;
+  FormID : TwbFormID;
 begin
   Result := nil;
+  FormID := TwbFormID.FromCardinal(aInt);
+
   if Assigned(aElement) then begin
     _File := aElement._File;
-    if Assigned(_File) then
-      Result := _File.RecordByFormID[TwbFormID.FromCardinal(aInt), True, aElement.MastersUpdated];
+    if Assigned(_File) then begin
+      if (FormID.ObjectID < $800) and not _File.AllowHardcodedRangeUse then begin
+        FormID.FileID := TwbFileID.Null;
+        Result := wbGetGameMasterFile.RecordByFormID[FormID, True, aElement.MastersUpdated];
+      end else
+        Result := _File.RecordByFormID[FormID, True, aElement.MastersUpdated];
+    end;
   end;
 end;
 
@@ -13410,10 +13439,24 @@ begin
 end;
 
 function TwbFormIDDefFormater.MastersUpdated(aInt: Int64; const aOld, aNew: TwbFileIDs; aOldCount, aNewCount: Byte; const aElement: IwbElement): Int64;
+var
+  _File: IwbFile;
+  AllowHardcodedRangeUse : Boolean;
 begin
   Result := aInt;
-  if (aInt < $800) or (aInt = $FFFFFFFF) and (IsValid('ACVA') or IsValid('FFFF')) then
+  if (aInt = $FFFFFFFF) and (IsValid('ACVA') or IsValid('FFFF')) then
     Exit;
+
+  AllowHardcodedRangeUse := False;
+  if Assigned(aElement) then begin
+    _File := aElement._File;
+    if Assigned(_File) then
+      AllowHardcodedRangeUse := _File.AllowHardcodedRangeUse;
+  end;
+
+  if not AllowHardcodedRangeUse then
+    if (aInt and $FFFFFF) < $800 then
+      Exit(aInt and $FFF);
 
   if aInt <> 0 then
     Result := FixupFormID(TwbFormID.FromCardinal(aInt), aOld, aNew, aOldCount, aNewCount).ToCardinal;
@@ -13559,9 +13602,18 @@ begin
     _File := aElement._File;
     if Assigned(_File) then begin
       try
-        if wbDisplayLoadOrderFormID then
-          FormID := _File.FileFormIDtoLoadOrderFormID(FormID, aElement.MastersUpdated);
-        MainRecord := _File.RecordByFormID[TwbFormID.FromCardinal(aInt), True, aElement.MastersUpdated];
+        if (FormID.ObjectID < $800) and not _File.AllowHardcodedRangeUse then begin
+          FormID.FileID := TwbFileID.Null;
+          MainRecord := wbGetGameMasterFile.RecordByFormID[FormID, True, aElement.MastersUpdated];
+        end else  begin
+          MainRecord := _File.RecordByFormID[FormID, True, aElement.MastersUpdated];
+          if wbDisplayLoadOrderFormID then
+            if Assigned(MainRecord) then
+              FormID := MainRecord.LoadOrderFormID
+            else
+              FormID := _File.FileFormIDtoLoadOrderFormID(FormID, aElement.MastersUpdated);
+        end;
+
         if Assigned(MainRecord) then begin
           Result := MainRecord.Name;
           if wbReportMode then
@@ -14771,7 +14823,19 @@ begin
     _File := aElement._File;
     if Assigned(_File) then begin
       try
-        MainRecord := _File.RecordByFormID[TwbFormID.FromCardinal(aInt), True, aElement.MastersUpdated];
+
+        if (FormID.ObjectID < $800) and not _File.AllowHardcodedRangeUse then begin
+          FormID.FileID := TwbFileID.Null;
+          MainRecord := wbGetGameMasterFile.RecordByFormID[FormID, True, False];
+        end else  begin
+          MainRecord := _File.RecordByFormID[FormID, True, aElement.MastersUpdated];
+          if wbDisplayLoadOrderFormID then
+            if Assigned(MainRecord) then
+              FormID := MainRecord.LoadOrderFormID
+            else
+              FormID := _File.FileFormIDtoLoadOrderFormID(FormID, aElement.MastersUpdated);
+        end;
+
         if Assigned(MainRecord) then begin
           Found := MainRecord.Signature;
           if fidcValidRefs.IndexOf(Found) < 0 then
@@ -14787,8 +14851,6 @@ begin
           end;
           Exit;
         end;
-        if wbDisplayLoadOrderFormID then
-          FormID := _File.FileFormIDtoLoadOrderFormID(FormID, aElement.MastersUpdated);
       except
         on E: Exception do begin
           Result := E.Message;